Refactor tests and utilities: streamline test node creation, enhance joy topic assertions, and remove unused utility functions.
