+++ ros2_ws/src/rcdt_panther/test/integration/test_core.py
@@ -11,10 +11,13 @@ from geometry_msgs.msg import Twist
 from launch import LaunchDescription
 from launch.actions import IncludeLaunchDescription
 from rcdt_utilities.launch_utils import assert_for_message
-from rcdt_utilities.test_utils import call_trigger_service, get_joint_position
+from rcdt_utilities.test_utils import (
+    call_trigger_service,
+    get_joint_position,
+    wait_until_active,
+)
 from rclpy.node import Node
 from sensor_msgs.msg import JointState
-from utils import EndToEndUtils
 
 
 @launch_pytest.fixture(scope="module")
@@ -35,13 +38,13 @@ def test_joint_states_published() -> None:
     assert_for_message(JointState, "/panther/joint_states", 60)
 
 
-def test_ready_to_start(singleton_node: Node) -> None:
+def test_ready_to_start(test_node: Node) -> None:
     """This test will ensure the tests are ready to start by waiting for the move_group and moveit_manager node.
     Also waits until the gripper_action_controller is active."""
 
     assert (
-        EndToEndUtils().wait_until_active(
-            node=singleton_node,
+        wait_until_active(
+            node=test_node,
             controller_name="drive_controller",
             controller_manager_name="/panther/controller_manager",
         )
@@ -50,25 +53,25 @@ def test_ready_to_start(singleton_node: Node) -> None:
 
 
 @pytest.mark.launch(fixture=core)
-def test_e_stop_request(singleton_node: Node) -> None:
+def test_e_stop_request(test_node: Node) -> None:
     assert (
         call_trigger_service(
-            node=singleton_node, service_name="/panther/hardware/e_stop_reset"
+            node=test_node, service_name="/panther/hardware/e_stop_reset"
         )
         is True
     )
 
 
 @pytest.mark.launch(fixture=core)
-def test_driving(singleton_node: Node) -> None:
+def test_driving(test_node: Node) -> None:
     """Test that the controllers work and the wheels have turned."""
-    pub = singleton_node.create_publisher(Twist, "/panther/cmd_vel", 10)
+    pub = test_node.create_publisher(Twist, "/panther/cmd_vel", 10)
 
     msg = Twist()
     msg.linear.x = 1.0
 
     pub.publish(msg)
-    rclpy.spin_once(singleton_node, timeout_sec=0.1)
+    rclpy.spin_once(test_node, timeout_sec=0.1)
     time.sleep(1)  # give the panther some time to move
 
     joint_value = get_joint_position("panther", "fl_wheel_joint")
