+++ ros2_ws/src/rcdt_utilities/rcdt_utilities/test_utils.py
@@ -3,24 +3,38 @@
 # SPDX-License-Identifier: Apache-2.0
 
 
-from typing import Iterator, Type
+import time
+from typing import Callable, Iterator, Type
 
 import pytest
-from typing import Type
 import rclpy
+from controller_manager_msgs.msg import ControllerState
+from controller_manager_msgs.srv import ListControllers
+from geometry_msgs.msg import PoseStamped
 from launch_testing_ros.wait_for_topics import WaitForTopics
+from rcdt_messages.srv import ExpressPoseInOtherFrame
 from rclpy.action import ActionClient
 from rclpy.client import Client
 from rclpy.node import Node
-from rclpy.service import Service
-from sensor_msgs.msg import JointState
-from std_srvs.srv import Trigger
-import time
-
-from rclpy.node import Node
 from rclpy.qos import QoSDurabilityPolicy, QoSHistoryPolicy, QoSProfile
-from sensor_msgs.msg import Joy
+from rclpy.service import Service
+from rclpy.task import Future
+from sensor_msgs.msg import JointState, Joy
 from std_msgs.msg import String
+from std_srvs.srv import Trigger
+
+
+@pytest.fixture(scope="session")
+def test_node() -> Iterator[Node]:
+    """Fixture to create a singleton node for testing."""
+    print("initialization is done")
+    if not rclpy.ok():
+        rclpy.init()
+    node = Node("test_node")
+    yield node
+    node.destroy_node()
+    if rclpy.ok():
+        rclpy.shutdown()
 
 
 def get_joint_position(namespace: str, joint: str) -> float:
@@ -102,21 +116,11 @@ def create_ready_action_client(
     return client
 
 
-@pytest.fixture(scope="session")
-def singleton_node() -> Iterator[Node]:
-    """Fixture to create a singleton node for testing."""
-    rclpy.init()
-    node = Node("test_node")
-    yield node
-    node.destroy_node()
-    rclpy.shutdown()
-
-
 def assert_joy_topic_switch(
     node: Node,
     expected_topic: str,
     button_config: list[int],
-    timeout_sec: float = 3.0,
+    timeout_sec: float = 10.0,
     state_topic: str = "/joy_topic_manager/state",
 ) -> None:
     """
@@ -153,10 +157,206 @@ def assert_joy_topic_switch(
     pub.publish(msg)
 
     start_time = time.time()
-    while "state" not in result and time.time() - start_time < timeout_sec:
+    while time.time() - start_time < timeout_sec:
         rclpy.spin_once(node, timeout_sec=0.1)
+        if result.get("state") == expected_topic:
+            break
         time.sleep(0.1)
 
     assert result.get("state") == expected_topic, (
         f"Expected state '{expected_topic}', but got '{result.get('state')}'"
     )
+
+
+def call_express_pose_in_other_frame(
+    node: Node, pose: PoseStamped, target_frame: str, timeout_sec: float = 5.0
+) -> ExpressPoseInOtherFrame.Response:
+    """
+    Calls the /express_pose_in_other_frame service.
+
+    Args:
+        node (Node): An active rclpy Node.
+        pose (PoseStamped): The pose to transform.
+        target_frame (str): The frame to express the pose in.
+        timeout_sec (float): Timeout for waiting on service and result.
+
+    Returns:
+        ExpressPoseInOtherFrame.Response: The response containing the transformed pose.
+    """
+
+    client = create_ready_service_client(
+        node, ExpressPoseInOtherFrame, "/express_pose_in_other_frame"
+    )
+
+    request = ExpressPoseInOtherFrame.Request()
+    request.pose = pose
+    request.target_frame = target_frame
+
+    future: Future = client.call_async(request)
+    rclpy.spin_until_future_complete(node, future, timeout_sec=timeout_sec)
+
+    response = future.result()
+    if response is None:
+        raise RuntimeError("Service call failed or timed out")
+
+    return response
+
+
+def assert_movements_with_joy(  # noqa: PLR0913
+    node: Node,
+    joy_axes: list[float],
+    compare_fn: Callable[[PoseStamped, PoseStamped], float],
+    threshold: float,
+    description: str,
+    frame_base: str = "panther/base_link",
+    frame_target: str = "panther/odom",
+) -> None:
+    """Publishes a joystick message and asserts that movement occurs above a threshold.
+
+    Args:
+        test_node: The ROS node used for publishing/subscribing.
+        joy_axes: A list of axes for the Joy message.
+        compare_fn: A function taking (first_pose, moved_pose) and returning a float delta.
+        threshold: The minimum delta required to pass the test.
+        description: Description of what is being tested (used in assertion message).
+        frame_base: The frame in which the original pose is defined.
+        frame_target: The frame to which the pose should be transformed.
+    """
+    pose = PoseStamped()
+    pose.header.frame_id = frame_base
+    first_pose = call_express_pose_in_other_frame(
+        node=node, pose=pose, target_frame=frame_target
+    ).pose.pose
+
+    pub = node.create_publisher(Joy, "/joy", 10)
+
+    msg = Joy()
+    msg.axes = joy_axes
+    pub.publish(msg)
+    rclpy.spin_once(node, timeout_sec=0.1)
+    time.sleep(1)  # let the robot move
+
+    pose = PoseStamped()
+    pose.header.frame_id = frame_base
+    moved_pose = call_express_pose_in_other_frame(
+        node=node, pose=pose, target_frame=frame_target
+    ).pose.pose
+    delta = compare_fn(first_pose, moved_pose)
+
+    assert abs(delta) > threshold, (
+        f"{description} did not change after input. Δ = {delta:.4f}"
+    )
+
+
+def list_controllers(
+    node: Node, controller_manager_name: str = "/franka/controller_manager"
+) -> list[ControllerState]:
+    """Query the controller manager for all currently loaded controllers.
+
+    Args:
+        node (Node): The rclpy node used to create the service client.
+        controller_manager_name (str): Name or namespace of the controller manager.
+
+    Returns:
+        List[ControllerState]: List of current controller states."""
+    client = create_ready_service_client(
+        node, ListControllers, f"{controller_manager_name}/list_controllers"
+    )
+    request = ListControllers.Request()
+    future: Future = client.call_async(request)
+    rclpy.spin_until_future_complete(node, future)
+
+    response: ListControllers.Response = future.result()
+    if response is None:
+        raise RuntimeError("Failed to get response from list_controllers")
+
+    return response.controller
+
+
+def get_controller_state(
+    controllers: list[ControllerState], controller_name: str
+) -> str:
+    """Retrieve the state of a controller by name.
+
+    Args:
+        controllers (List[ControllerState]): List of controllers.
+        controller_name (str): Name of the controller to find.
+
+    Returns:
+        str: The current state of the controller.
+
+    Raises:
+        ValueError: If the controller is not found."""
+    for controller in controllers:
+        if controller.name == controller_name:
+            return controller.state
+    raise ValueError(f"Controller '{controller_name}' not found")
+
+
+def wait_until_active(
+    node: Node,
+    controller_name: str,
+    timeout_sec: float = 90.0,
+    poll_interval: float = 0.5,
+    controller_manager_name: str = "/franka/controller_manager",
+) -> bool:
+    """Poll until a controller becomes 'active'.
+
+    Args:
+        node (Node): rclpy node used to call the service.
+        controller_name (str): Name of the controller to wait for.
+        timeout_sec (float): Timeout duration in seconds.
+        poll_interval (float): Interval between polls.
+        controller_manager_name (str): Controller manager service name.
+
+    Returns:
+        bool: True if the controller became active, False on timeout."""
+    end_time = time.time() + timeout_sec
+    while time.time() < end_time:
+        controllers = list_controllers(node, controller_manager_name)
+        try:
+            state = get_controller_state(controllers, controller_name)
+            if state == "active":
+                print("The controller is active!")
+                time.sleep(1)
+                return True
+        except ValueError:
+            pass
+
+        time.sleep(poll_interval)
+
+    return False
+
+
+def wait_until_reached_joint(
+    namespace: str,
+    joint: str,
+    expected_value: float,
+    tolerance: float = 0.025,
+    timeout_sec: int = 30,
+) -> tuple[bool, float]:
+    """Wait until a joint reaches the expected value within a tolerance.
+
+    Args:
+        namespace (str): Namespace of the robot (e.g., 'franka').
+        joint (str): Name of the joint to check.
+        expected_value (float): Target joint value in radians.
+        tolerance (float): Acceptable deviation from the expected value.
+        timeout_sec (int): Timeout duration in seconds.
+        poll_interval (float): Interval between joint state checks.
+
+    Returns:
+        Tuple[bool, float]: (True, joint_value) if target reached; otherwise (False"""
+    end_time = time.time() + timeout_sec
+    while time.time() < end_time:
+        try:
+            joint_value = get_joint_position(namespace=namespace, joint=joint)
+            if joint_value == pytest.approx(expected_value, abs=tolerance):
+                print("The joint reached its expected value!")
+                time.sleep(0.25)
+                return (True, joint_value)
+        except ValueError:
+            pass
+
+        time.sleep(0.25)
+    return (False, joint_value)
